OnlineStore.WebServer/requirements.txt
Django==2.2.1
django-cors-headers==3.0.2
django-rest-auth==0.9.5
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
PyJWT==1.7.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
